pyproject.toml
[project]
name = "basyx-opcua-bridge"
version = "1.0.0"
description = "High-performance OPC UA to AAS Synchronizer"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "basyx-python-sdk>=1.0.0",
    "asyncua>=1.1.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "anyio>=4.3.0",
    "typer>=0.9.0",
    "orjson>=3.9.0",
    "prometheus-client>=0.20.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0"
]

[project.scripts]
basyx-bridge = "basyx_opcua_bridge.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
